Split up the scan code into multiple source files
